image service/content handler improvements
image service: use generic ImageResult and a generic imageHanderFunc that a
caller passes in to produce an artifact from the image. this reduces the
amount of intermediate caching of image data that is required.

content handler: allow the content to be served from an arbitrary URL
base path (as provided via a CLI flag)

Signed-off-by: Joe Lanford <joe.lanford@gmail.com>